Fixed bug in calculate_base_resistance
If the pile dimeter is not a multiple of 0.2m, the unit base resistance is calculated for the closeby diameters which are a multiple of 0.2. 

Before the bug fix, a diameter of 0.21 would result in a calculation for round(0.21 - 0.01 ( = 0.21 % 0.2 ), 1) = 0.20 and round(0.21 + 0.01 ( = 0.21 % 0.2 ) , 1) = 0.20. This is a mistake, it should be 0.20 (lower bound) and 0.25 (upper bound).